Fix loading of config.options.enabledEnvironments
Today a `config.options.enabledEnvironments` can be provided to allow generating docs in these envs. This is documented in README. However, `postprocessTree` method is looking for `config.enabledEnvironments`. Result is that `ember ember-cli-yuidoc` is working well but `ember build` never generates docs, even if the current environment is provided in conf. This fixes it.